pyproject.toml
[project]
name = "prynai-mcp"
version = "0.1.0"
requires-python = ">=3.10"
dependencies = [
  "httpx>=0.27",
  "mcp[cli]>=1.9.0", # official SDK with FastMCP + Streamable HTTP
  "msal>=1.28",
  "pydantic-settings>=2.3",
  "pyjwt[crypto]>=2.8",
  "redis>=5.0.4",
  "starlette>=0.37",
  "uvicorn>=0.30",
]

[project.scripts]
prynai-mcp = "prynai_mcp.server:main"
